portable: use C99 PRIu64/PRId64/PRIx64 for fixed-width format strings
Replace all %llu/%lld/%llx/%#llx with the PRI macros from <inttypes.h>,
which resolve to correct format specifiers on both macOS (llu/lld/llx)
and Linux (lu/ld/lx). Drop the (unsigned long long) casts in favor
of (uint64)/(int64) to match the PRI argument types.

+++ apps/cli/test.c
@@ -166,7 +166,7 @@ bitc_crypt_test(void)
    s = crypt_set_key_from_passphrase(pass, &k, &count);
    ASSERT(s);
 
-   printf("num_iterations = %lld\n", (long long)count);
+   printf("num_iterations = %" PRId64 "\n", (int64)count);
 
    s = crypt_encrypt(&k, sec, &cipher, &clen);
    ASSERT(s);
